move subgraph clients
