+++ constants/config.py
@@ -104,3 +104,5 @@ TOKEN_SYMBOL_MAP = {
     BEAN_ETH_WELL_ADDR.lower(): "BEANETH",
     BEAN_WSTETH_WELL_ADDR.lower(): "BEANwstETH",
 }
+
+GRAPH_FIELDS_PLACEHOLDER = "_FIELDS_"
