+++ monitors/curve.py
@@ -3,7 +3,6 @@ from data_access.coin_gecko import get_token_price
 from bots.util import *
 from monitors.monitor import Monitor
 from data_access.eth_chain import *
-from data_access.graphs import *
 from data_access.util import *
 from constants.addresses import *
 from constants.config import *
