+++ monitors/preview/eth.py
@@ -4,7 +4,6 @@ from data_access.eth_usd_oracle import get_twa_eth_price
 from bots.util import *
 from monitors.preview.preview import PreviewMonitor
 from data_access.eth_chain import *
-from data_access.graphs import *
 from data_access.util import *
 from constants.addresses import *
 from constants.config import *
