+++ monitors/preview/nft.py
@@ -3,7 +3,6 @@ from opensea import OpenseaAPI
 from bots.util import *
 from monitors.preview.preview import PreviewMonitor
 from data_access.eth_chain import *
-from data_access.graphs import *
 from data_access.util import *
 from constants.addresses import *
 from constants.config import *
