+++ monitors/preview/preview.py
@@ -1,7 +1,6 @@
 from bots.util import *
 from monitors.monitor import Monitor
 from data_access.eth_chain import *
-from data_access.graphs import *
 from data_access.util import *
 from constants.addresses import *
 from constants.config import *
