+++ monitors/preview/price.py
@@ -1,7 +1,7 @@
 from bots.util import *
 from monitors.preview.preview import PreviewMonitor
 from data_access.eth_chain import *
-from data_access.graphs import *
+from data_access.subgraphs.beanstalk import BeanstalkGraphClient
 from data_access.util import *
 from constants.addresses import *
 from constants.config import *
@@ -18,7 +18,7 @@ class PricePreviewMonitor(PreviewMonitor):
 
     def _monitor_method(self):
         self.bean_client = BeanClient()
-        self.beanstalk_graph_client = BeanstalkSqlClient()
+        self.beanstalk_graph_client = BeanstalkGraphClient()
         while self._thread_active:
             self.wait_for_next_cycle()
             self.iterate_display_index()
